refactor(desktop): connections are a hook
