+++ apps/desktop/src/renderer/app.tsx
@@ -4,10 +4,7 @@ import {
   PRODUCT_SEARCH_SURFACE,
   PRODUCT_SURFACE_EVENT,
 } from "@sidecar/analytics";
-import { APPLE_CALENDAR_ACCESS, APPLE_CALENDAR_ID } from "@sidecar/calendar/vocabulary";
-import type { AccountProvider } from "@sidecar/credentials/snapshot";
 import { ACCOUNT_STATUS } from "@sidecar/credentials/snapshot";
-import type { CredentialProviderId } from "@sidecar/credentials/vocabulary";
 import {
   CREDENTIAL_PROVIDER_LIST,
   CREDENTIAL_PROVIDERS,
@@ -25,11 +22,7 @@ import {
 } from "@sidecar/session";
 import { FIXTURE_EPOCH_MS, FIXTURE_SPEAKING_CAPTION } from "@sidecar/session/fixtures";
 import { APP_SETTING_SCHEMA, VOICE_HOTKEY_NONE, voiceHotkeyLabel } from "@sidecar/settings";
-import type {
-  AppSettingsView,
-  ObservedAccountCalendars,
-  SettingsUpdateResult,
-} from "@sidecar/settings/wire";
+import type { AppSettingsView, ObservedAccountCalendars } from "@sidecar/settings/wire";
 import { appSettingsView } from "@sidecar/settings/wire";
 import { MOTION_DURATION_MS, VOICE_CAPTION_MAX_HEIGHT } from "@sidecar/surface";
 import { cssCustomProperties } from "@sidecar/surface/react-css";
@@ -43,7 +36,6 @@ import {
   useRef,
   useState,
 } from "react";
-import { CONSENT_SERVICE_ID, type ConsentServiceId } from "#shared/consent-services";
 import { ACT_KIND } from "#shared/messages/acts";
 import { type AppStateSnapshot, sessionReplayBootstrap } from "#shared/messages/app-state";
 import type {
@@ -59,9 +51,7 @@ import {
 import { act, tell, updateSetting, updateSettingEntry } from "./act";
 import { ASK_LUKE_INPUT_ID, focusAskField } from "./ask-luke";
 import type { CalendarGateControl } from "./calendar-gate";
-import { type ConsentConnectEntry, ConsentConnectSlot } from "./consent-connect-slot";
-import type { CredentialEntry, CredentialEntryControl } from "./credential-entry";
-import { isSubmittable, removalEndsEntry } from "./credential-entry";
+import { ConsentConnectSlot } from "./consent-connect-slot";
 import {
   armErrand,
   EMPTY_ERRAND_RUN,
@@ -108,12 +98,7 @@ import { parsePixels } from "./session-motion";
 import { SESSION_OPTIONS_CONTROL_ID, SESSION_OPTIONS_ID } from "./session-parts";
 import { applySessionReplay } from "./session-replay";
 import { focusSearchField, SESSION_SEARCH_INPUT_ID } from "./session-search";
-import type {
-  MicrophoneControl,
-  ShortcutControl,
-  SupersetControl,
-  UpdateControl,
-} from "./settings-panel";
+import type { MicrophoneControl, ShortcutControl, UpdateControl } from "./settings-panel";
 import { SETTINGS_SEARCH_INPUT_ID } from "./settings-search";
 import {
   credentialSettingsPage,
@@ -121,8 +106,6 @@ import {
   SETTING_PAGE,
   SETTINGS_VIEW,
   type SettingsView,
-  type SlotOccupant,
-  standDownReturnPage,
 } from "./settings-views";
 import { useSignInFaceCycle } from "./sign-in-gate";
 import { SignInSlot } from "./sign-in-slot";
@@ -136,9 +119,10 @@ import {
 import { SupersetSignInSlot } from "./superset-sign-in-slot";
 import { UPDATE_ROW_ACTION, updateRow } from "./update-row";
 import { appSettingsNow, appStateNow, useAppState } from "./use-app-state";
+import { useConnections } from "./use-connections";
 import { useFeedbackComposer } from "./use-feedback-composer";
 import { useMeasuredHeight } from "./use-measured-height";
-import { type PanelEntrySurface, panelEntryOpen, usePanelEntry } from "./use-panel-entry";
+import type { PanelEntrySurface } from "./use-panel-entry";
 import { usePanelPresentation } from "./use-panel-presentation";
 import { usePrefersReducedMotion } from "./use-reduced-motion";
 import { useStateWithRef } from "./use-state-with-ref";
@@ -159,35 +143,6 @@ import {
  * keystrokes, short enough that quitting mid-thought still keeps the search.
  */
 const SEARCH_QUERY_STORE_DELAY_MS = 400;
-
-/**
- * The bridge acts behind each consent service's wait: the documented connect
- * the slot's send runs, the main-process side a mid-wait cancel must stop,
- * and — for the browser flows alone — the way a lost tab reopens. One row
- * per service, so a fourth service is a fourth row rather than a fourth
- * branch at every dispatch site.
- */
-const CONSENT_ACTIONS = {
-  [CONSENT_SERVICE_ID.APPLE_CALENDAR]: {
-    connect: () => act(ACT_KIND.CALENDAR_CONNECT_APPLE),
-    cancel: () => tell(ACT_KIND.CALENDAR_CANCEL_APPLE_CONNECT),
-  },
-  [CONSENT_SERVICE_ID.GOOGLE_CALENDAR]: {
-    connect: () => act(ACT_KIND.CALENDAR_CONNECT_GOOGLE),
-    cancel: () => tell(ACT_KIND.CALENDAR_CANCEL_GOOGLE_SIGN_IN),
-    reopen: () => tell(ACT_KIND.CALENDAR_REOPEN_GOOGLE_SIGN_IN),
-  },
-  [CONSENT_SERVICE_ID.LINEAR]: {
-    connect: () => act(ACT_KIND.TRACKER_CONNECT),
-    cancel: () => tell(ACT_KIND.TRACKER_CANCEL_SIGN_IN),
-    reopen: () => tell(ACT_KIND.TRACKER_REOPEN_SIGN_IN),
-  },
-} as const satisfies Readonly<
-  Record<
-    ConsentServiceId,
-    { connect: () => Promise<SettingsUpdateResult>; cancel: () => void; reopen?: () => void }
-  >
->;
 
 /**
  * The composer kind a spoken open names, matched to the composer's own. The
@@ -398,11 +353,6 @@ export function App(): React.JSX.Element {
   /** Whether a calendar sign-in holds the slot, mirrored like the other two. */
   const consentConnectHeld = useRef(false);
   const supersetSignInHeld = useRef(false);
-  /**
-   * Which entry the slot shape is drawn around — a key being pasted or either
-   * sign-in being waited out. One shape, three occupants, never together.
-   */
-  const slotOccupant = useRef<SlotOccupant>(PANEL_STAND_DOWN.KEY);
   /**
    * How many errands Luke has run. Carried with each one so that asking for
    * the same control twice flies twice, exactly as a repeated face gesture is
@@ -867,322 +817,23 @@ export function App(): React.JSX.Element {
     return () => window.clearTimeout(settled);
   }, [state, sessionView.query, liveSettings?.sessionSearchQuery]);
 
-  const removeCalendarAccount = useCallback(
-    (accountId: string) => act(ACT_KIND.CALENDAR_REMOVE_ACCOUNT, { accountId }),
-    [],
-  );
-
-  const toggleCalendarSelected = useCallback(
-    (accountId: string, calendarId: string, selected: boolean) =>
-      act(ACT_KIND.CALENDAR_SET_SELECTED, { accountId, calendarId, selected }),
-    [],
-  );
-
-  const disconnectAppleCalendar = useCallback(() => act(ACT_KIND.CALENDAR_DISCONNECT_APPLE), []);
-
-  const toggleAppleCalendarSelected = useCallback(
-    (calendarId: string, selected: boolean) =>
-      act(ACT_KIND.CALENDAR_SET_SELECTED, { accountId: APPLE_CALENDAR_ID, calendarId, selected }),
-    [],
-  );
-
-  const disconnectLinear = useCallback(() => act(ACT_KIND.TRACKER_DISCONNECT), []);
-
-  /**
-   * A consent sign-in is asking for one thing too, so the panel gets out of
-   * the way of it the same way it does for a key: the shape goes down to a
-   * slot that says what it is waiting for. The flow itself runs in the
-   * browser and the main process; when the grant lands, the panel comes back
-   * around the newly connected service. One entry serves every such sign-in,
-   * because only one consent page can be open at a time and there is only one
-   * slot for it to stand in.
-   */
-  const consentConnect = usePanelEntry<ConsentConnectEntry>({
-    aside: PANEL_PRESENTATION.SLOT,
-    // Giving up mid-wait leaves — the consent page is where the user is — but
-    // a sign-in that failed is read in the slot, so its Close restores the
-    // panel to try again from the row.
-    restoresPanel: (held) => held.rejection !== undefined,
-    isSendable: (entry): entry is ConsentConnectEntry => entry !== undefined && !entry.busy,
-    send: async (sending) => {
-      // Which service is being connected decides which documented act runs,
-      // and nothing else does: the entry names a row of the actions table.
-      const result = await CONSENT_ACTIONS[sending.serviceId].connect();
-      return result.reason ? { rejection: result.reason } : {};
+  const connections = useConnections({
+    surface: panelEntrySurface,
+    credentialHeld,
+    consentConnectHeld,
+    supersetSignInHeld,
+    rememberStandDownPage,
+    expand,
+    calendars,
+    superset: {
+      installed: state?.superset.installed === true,
+      connected: supersetConnected,
+      signIn: supersetSignIn,
+      defaultAgent: settings?.workspaceAgentDefaults?.[SUPERSET_WORKSPACE_PROVIDER_ID]?.agent,
     },
-    pointerInside: pointerIsInside,
-    presentation: presentationOf,
-    onReleasedWhileAway: onHitRegionLeave,
-    cancelHover,
-    applyPresentation,
-    restorePanel,
-    leave,
-    settle,
-    heldRef: consentConnectHeld,
+    workspaceProjects,
   });
-
-  /** One press: stand down to the waiting slot and open the consent page. */
-  const beginConsentSignIn = useCallback(
-    (serviceId: ConsentServiceId) => {
-      // One slot, one occupant: a key mid-paste is not disturbed by a
-      // sign-in, and neither is another sign-in already waiting.
-      if (credentialHeld.current || consentConnectHeld.current) return;
-      slotOccupant.current = PANEL_STAND_DOWN.CONSENT;
-      // Every consent block stands under Integrations, so that is where a
-      // cancelled or refused sign-in comes back to.
-      standDownPage.current = standDownReturnPage({ kind: PANEL_STAND_DOWN.CONSENT });
-      consentConnect.begin({ serviceId, busy: false });
-      consentConnect.commit();
-    },
-    [consentConnect.begin, consentConnect.commit],
-  );
-
-  /** True while a granted-already connect runs on the row, dialog-free. */
-  const [appleCalendarBusy, setAppleCalendarBusy] = useState(false);
-
-  /**
-   * Connecting this Mac's Calendar stands the panel down only when macOS is
-   * actually about to ask: with the grant already standing, the dialog never
-   * appears, and a stand-down would flash the slot for a single frame — so
-   * the press probes the status first and connects in place when it can.
-   */
-  const connectAppleCalendar = useCallback(async () => {
-    setAppleCalendarBusy(true);
-    let granted = false;
-    try {
-      granted = (await act(ACT_KIND.CALENDAR_APPLE_ACCESS_STATUS)) === APPLE_CALENDAR_ACCESS.FULL;
-      if (granted) await act(ACT_KIND.CALENDAR_CONNECT_APPLE);
-    } finally {
-      setAppleCalendarBusy(false);
-    }
-    if (!granted) beginConsentSignIn(CONSENT_SERVICE_ID.APPLE_CALENDAR);
-  }, [beginConsentSignIn]);
-
-  /** The Mac's own entry in the latest calendars broadcast, for its block. */
-  const appleCalendarObserved = calendars.find((choice) => choice.accountId === APPLE_CALENDAR_ID);
-
-  /**
-   * Whether another entry holds the one slot, which refuses this service's
-   * Connect: a key mid-paste, or a different service's consent wait.
-   */
-  const slotHeldExcept = (serviceId: ConsentServiceId) =>
-    credentialsEntry.entry !== undefined ||
-    (consentConnect.entry !== undefined && consentConnect.entry.serviceId !== serviceId);
-
-  const cancelConsentSignIn = useCallback(() => {
-    // Mid-wait, the flow's main-process side must stop listening too — a
-    // browser flow's loopback, or Apple's System Settings watch; after a
-    // failure there is nothing left to stop.
-    const waiting = consentConnect.latest();
-    if (waiting?.busy) CONSENT_ACTIONS[waiting.serviceId].cancel();
-    consentConnect.cancel();
-  }, [consentConnect.cancel, consentConnect.latest]);
-
-  // The stage the slot draws is the document's throughout: the flow moves it
-  // as it goes — browser code, exchanging, the organization choice — and each
-  // stage reaches every window before the press's own reply returns.
-  const beginSupersetSignIn = useCallback(() => {
-    if (supersetSignInHeld.current) {
-      if (supersetSignIn.stage === SUPERSET_SIGN_IN_STAGE.FAILURE) {
-        tell(ACT_KIND.SUPERSET_BEGIN_SIGN_IN);
-      }
-      return;
-    }
-    if (credentialHeld.current || consentConnectHeld.current) return;
-    supersetSignInHeld.current = true;
-    slotOccupant.current = PANEL_STAND_DOWN.SUPERSET;
-    standDownPage.current = standDownReturnPage({ kind: PANEL_STAND_DOWN.SUPERSET });
-    cancelHover();
-    applyPresentation(PANEL_PRESENTATION.SLOT);
-    tell(ACT_KIND.SUPERSET_BEGIN_SIGN_IN);
-  }, [applyPresentation, cancelHover, supersetSignIn.stage]);
-
-  const cancelSupersetSignIn = useCallback(() => {
-    if (!supersetSignInHeld.current) return;
-    tell(ACT_KIND.SUPERSET_CANCEL_SIGN_IN);
-    supersetSignInHeld.current = false;
-    if (presentationOf() === PANEL_PRESENTATION.SLOT) restorePanel();
-  }, [presentationOf, restorePanel]);
-
-  const disconnectSuperset = useCallback(() => act(ACT_KIND.SUPERSET_DISCONNECT), []);
-
-  /**
-   * Asking to write a key is asking for one thing, so the panel gets out of the
-   * way of it: the shape goes down to the slot, which is the field and nothing
-   * else. It is the same wherever the key is coming from — a first connection, a
-   * stored key being replaced, or one standing in front of the environment's —
-   * because they are all the same action.
-   */
-  const credentialsEntry = usePanelEntry<CredentialEntry>({
-    aside: PANEL_PRESENTATION.SLOT,
-    restoresPanel: (held) => held.away !== true,
-    isSendable: isSubmittable,
-    send: async (sending) => {
-      const result = await act(ACT_KIND.CREDENTIAL_SET_API_KEY, {
-        providerId: sending.providerId,
-        apiKey: sending.draft,
-      });
-      return result.reason ? { rejection: result.reason } : {};
-    },
-    pointerInside: pointerIsInside,
-    presentation: presentationOf,
-    onReleasedWhileAway: onHitRegionLeave,
-    cancelHover,
-    applyPresentation,
-    restorePanel,
-    leave,
-    settle,
-    heldRef: credentialHeld,
-  });
-
-  const beginEntry = useCallback(
-    (providerId: CredentialProviderId) => {
-      // Where the entry's row is drawn, remembered before the trip to the
-      // slot so coming back lands on the page the entry began on.
-      standDownPage.current = standDownReturnPage({ kind: PANEL_STAND_DOWN.KEY, providerId });
-      slotOccupant.current = PANEL_STAND_DOWN.KEY;
-      credentialsEntry.begin({ providerId, draft: "", busy: false, away: false });
-    },
-    [credentialsEntry.begin],
-  );
-
-  /**
-   * A Connect press: the same entry {@link beginEntry} opens, with the
-   * provider's key page opened in the same press. Someone connecting has no
-   * key yet, so the first thing they need is the page that issues one — the
-   * consent and CLI connectors already work this way, because their browser
-   * half *is* the connection. The entry starts `away` for the same reason
-   * {@link fetchKey} marks it: the person this slot is now waiting for is
-   * reading a browser, so giving up leaves the browser alone rather than
-   * bringing the panel back over it. The page is still opened by provider id,
-   * so the only addresses reachable are the ones the credential registry
-   * fixes.
-   */
-  const connectEntry = useCallback(
-    (providerId: CredentialProviderId) => {
-      standDownPage.current = standDownReturnPage({ kind: PANEL_STAND_DOWN.KEY, providerId });
-      slotOccupant.current = PANEL_STAND_DOWN.KEY;
-      tell(ACT_KIND.CREDENTIAL_OPEN_API_KEYS, { providerId });
-      credentialsEntry.begin({ providerId, draft: "", busy: false, away: true });
-    },
-    [credentialsEntry.begin],
-  );
-
-  /**
-   * Sends the browser to the provider's key page. The entry remembers that it
-   * did: from here on, the person this slot is waiting for is reading a page
-   * that Luke — which floats above every window — would otherwise be sitting on
-   * top of. It is also what the slot is for, so the shape is already right; the
-   * panel is only stood down if the link was pressed from inside it.
-   */
-  const fetchKey = useCallback(() => {
-    const current = credentialsEntry.latest();
-    // Same rule as typing: the key on its way to the store is what the entry is
-    // for, and going to fetch another one is not a reason to disturb it. Both
-    // views disable the link while it is in flight, so this is the floor rather
-    // than the answer.
-    if (!panelEntryOpen(current)) return;
-    tell(ACT_KIND.CREDENTIAL_OPEN_API_KEYS, { providerId: current.providerId });
-    credentialsEntry.apply({ ...current, away: true });
-    if (presentationOf() === PANEL_PRESENTATION.SLOT) return;
-    credentialsEntry.standDown();
-  }, [credentialsEntry.apply, credentialsEntry.latest, credentialsEntry.standDown, presentationOf]);
-
-  const removeProviderApiKey = useCallback(
-    async (providerId: CredentialProviderId) => {
-      const result = await act(ACT_KIND.CREDENTIAL_SET_API_KEY, { providerId, apiKey: undefined });
-      // Delete and the field are on the row together once the panel has been
-      // brought back around an entry, and a key that has been removed cannot be
-      // replaced.
-      if (removalEndsEntry(credentialsEntry.latest(), providerId, result.reason)) {
-        credentialsEntry.apply(undefined);
-      }
-      return result;
-    },
-    [credentialsEntry.apply, credentialsEntry.latest],
-  );
-
-  const credentials: CredentialEntryControl = {
-    entry: credentialsEntry.entry,
-    begin: beginEntry,
-    connect: connectEntry,
-    change: (draft) => credentialsEntry.patch({ draft }),
-    fetchKey,
-    cancel: credentialsEntry.cancel,
-    commit: credentialsEntry.commit,
-    remove: removeProviderApiKey,
-  };
-
-  /**
-   * Whose sign-in the surface is waiting on. Choosing a provider on the gate
-   * sends the browser to it and stands the panel down to a small waiting
-   * popup, the way fetching an API key stands it down to the slot: the real
-   * work is in the browser, and Luke floats above the page it needs. Held as
-   * app state with a ref because the attempt's own reply has to read whether
-   * it is still the one being waited on.
-   */
-  const [signInWait, setSignInWait, signInWaitNow] = useStateWithRef<AccountProvider | undefined>(
-    undefined,
-  );
-  /**
-   * Which attempt any reply answers. Cancel advances it, so the outcome of a
-   * sign-in already withdrawn is spent rather than moving the shape again.
-   */
-  const signInAttempt = useRef(0);
-  const [signInFailure, setSignInFailure] = useState<string>();
-
-  const beginSignIn = useCallback(
-    (provider: AccountProvider) => {
-      if (signInWaitNow() !== undefined) return;
-      const attempt = ++signInAttempt.current;
-      setSignInFailure(undefined);
-      setSignInWait(provider);
-      cancelHover();
-      applyPresentation(PANEL_PRESENTATION.SLOT);
-      act(ACT_KIND.ACCOUNT_BEGIN_SIGN_IN, { provider }).then(
-        () => {
-          if (signInAttempt.current !== attempt) return;
-          setSignInWait(undefined);
-          if (presentationOf() !== PANEL_PRESENTATION.SLOT) return;
-          // The panel comes forward around what was just unlocked — the
-          // session roster — and stays open like any other opened panel:
-          // signing in is an arrival, not an errand to settle and leave.
-          expand();
-        },
-        () => {
-          if (signInAttempt.current !== attempt) return;
-          setSignInWait(undefined);
-          setSignInFailure("Sign-in did not finish. Try again when you’re ready.");
-          if (presentationOf() === PANEL_PRESENTATION.SLOT) expand();
-        },
-      );
-    },
-    [applyPresentation, cancelHover, expand, presentationOf, setSignInWait, signInWaitNow],
-  );
-
-  /**
-   * Takes the wait back. The main process withdraws the loopback and signs the
-   * attempt back out; the panel returns to the gate at once rather than
-   * waiting for that round trip, and the attempt's eventual rejection finds
-   * itself already spent.
-   */
-  const cancelSignIn = useCallback(() => {
-    if (signInWaitNow() === undefined) return;
-    signInAttempt.current += 1;
-    setSignInWait(undefined);
-    tell(ACT_KIND.ACCOUNT_CANCEL_SIGN_IN);
-    if (presentationOf() === PANEL_PRESENTATION.SLOT) expand();
-  }, [expand, presentationOf, setSignInWait, signInWaitNow]);
-
-  const changeSupersetAgentDefault = useCallback(
-    (agent: string | undefined) =>
-      act(ACT_KIND.SETTING_UPDATE_ENTRY, {
-        field: APP_SETTING_SCHEMA.workspaceAgentDefaults.field,
-        key: SUPERSET_WORKSPACE_PROVIDER_ID,
-        value: agent === undefined ? undefined : { agent },
-      }),
-    [],
-  );
+  const slotOccupant = connections.slotOccupant;
 
   /**
    * The providers the default-workspace rows can offer: every provider
@@ -1932,10 +1583,17 @@ export function App(): React.JSX.Element {
       // here lands where it would have in the flow this is standing in for.
       changeTab(PANEL_TAB.SETTINGS);
       setSettingsView(credentialSettingsPage(firstProvider.id));
-      beginEntry(firstProvider.id);
+      connections.beginEntry(firstProvider.id);
     }
     window.sidecar.notifyReady();
-  }, [state, applyAuthoritativeMode, applyPresentation, beginEntry, changeTab, setSettingsView]);
+  }, [
+    state,
+    applyAuthoritativeMode,
+    applyPresentation,
+    connections.beginEntry,
+    changeTab,
+    setSettingsView,
+  ]);
 
   // The mode main decided, and the two events only a window can be told: the
   // ask field summoned from any app, and the composer a spoken request opens.
@@ -1971,17 +1629,6 @@ export function App(): React.JSX.Element {
     feedback.takeSpokenDraft,
     summonAsk,
   ]);
-
-  // A Superset sign-in carried through to the end gives the panel back around
-  // the newly connected service. Only the wait's own slot is answered: a
-  // connection the document reports for any other reason has no slot standing
-  // to come back from.
-  useEffect(() => {
-    if (!supersetSignInHeld.current) return;
-    if (supersetSignIn.stage !== SUPERSET_SIGN_IN_STAGE.CONNECTED) return;
-    supersetSignInHeld.current = false;
-    if (presentationOf() === PANEL_PRESENTATION.SLOT) restorePanel();
-  }, [supersetSignIn, presentationOf, restorePanel]);
 
   // The one greeting an unauthed launch gets: the panel opens on the sign-in
   // gate exactly once, then behaves like any panel — Escape, the pointer, and
@@ -2073,9 +1720,9 @@ export function App(): React.JSX.Element {
       // else it could mean. The sign-in wait and the consent connect borrow
       // the same shape, so the same key withdraws whichever is holding it.
       if (presentation === PANEL_PRESENTATION.SLOT) {
-        if (signInWaitNow() !== undefined) cancelSignIn();
-        else if (consentConnect.latest()) cancelConsentSignIn();
-        else credentialsEntry.cancel();
+        if (connections.signInWaitNow() !== undefined) connections.cancelSignIn();
+        else if (connections.consentWaiting()) connections.cancelConsentSignIn();
+        else connections.cancelEntry();
         return;
       }
       // Escape out of the composer leaves the shape and keeps the draft: a
@@ -2106,10 +1753,6 @@ export function App(): React.JSX.Element {
     window.addEventListener("keydown", handleKey);
     return () => window.removeEventListener("keydown", handleKey);
   }, [
-    cancelSignIn,
-    consentConnect.latest,
-    cancelConsentSignIn,
-    credentialsEntry.cancel,
     changeMode,
     changeTab,
     closeSearch,
@@ -2124,7 +1767,11 @@ export function App(): React.JSX.Element {
     settingsSearchOpen,
     setSettingsView,
     settingsView,
-    signInWaitNow,
+    connections.cancelConsentSignIn,
+    connections.cancelEntry,
+    connections.cancelSignIn,
+    connections.consentWaiting,
+    connections.signInWaitNow,
     speaking,
     stopSpeaking,
     tab,
@@ -2256,8 +1903,8 @@ export function App(): React.JSX.Element {
   // What the slot's field is for depends on what answers for that provider now,
   // and settings resolve after the first render.
   const slotSource =
-    credentialsEntry.entry && settings
-      ? settings.credentialSources[credentialsEntry.entry.providerId]
+    connections.credentialEntry && settings
+      ? settings.credentialSources[connections.credentialEntry.providerId]
       : CREDENTIAL_SOURCE.NONE;
   const microphone: MicrophoneControl = {
     status: state.audio.microphoneStatus,
@@ -2308,19 +1955,16 @@ export function App(): React.JSX.Element {
           ...(gateSettings.appleCalendarAvailable
             ? {
                 apple: {
-                  connecting:
-                    appleCalendarBusy ||
-                    consentConnect.entry?.serviceId === CONSENT_SERVICE_ID.APPLE_CALENDAR,
-                  onConnect: () => void connectAppleCalendar(),
+                  connecting: connections.appleCalendar.connecting,
+                  onConnect: connections.appleCalendar.onSignIn,
                 },
               }
             : undefined),
           ...(gateSettings.calendarSignInAvailable
             ? {
                 google: {
-                  connecting:
-                    consentConnect.entry?.serviceId === CONSENT_SERVICE_ID.GOOGLE_CALENDAR,
-                  onConnect: () => beginConsentSignIn(CONSENT_SERVICE_ID.GOOGLE_CALENDAR),
+                  connecting: connections.calendar.connecting,
+                  onConnect: connections.calendar.onSignIn,
                 },
               }
             : undefined),
@@ -2368,7 +2012,7 @@ export function App(): React.JSX.Element {
         // height of whichever is actually drawn.
         ...surfaceHeightStyle(
           panelHeight,
-          signInWait !== undefined
+          connections.signInWait !== undefined
             ? signInSlotHeight
             : slotOccupant.current === PANEL_STAND_DOWN.CONSENT ||
                 slotOccupant.current === PANEL_STAND_DOWN.SUPERSET
@@ -2391,8 +2035,10 @@ export function App(): React.JSX.Element {
           <PanelBody
             accountRequired={state.run.accountRequired}
             account={state.account}
-            onBeginSignIn={beginSignIn}
-            {...(signInFailure ? { signInFailure } : undefined)}
+            onBeginSignIn={connections.beginSignIn}
+            {...(connections.signInFailure
+              ? { signInFailure: connections.signInFailure }
+              : undefined)}
             {...(calendarGate ? { calendarGate } : undefined)}
             list={list}
             sessionsSettled={sessionsSettled}
@@ -2448,60 +2094,14 @@ export function App(): React.JSX.Element {
               microphone,
               updates,
               settings,
-              credentials,
+              credentials: connections.credentials,
               feedback: feedback.control,
               panelOpen,
               workspaceProviders: workspaceProviderOptions,
-              calendar: {
-                choices: calendars,
-                held: slotHeldExcept(CONSENT_SERVICE_ID.GOOGLE_CALENDAR),
-                connecting: consentConnect.entry?.serviceId === CONSENT_SERVICE_ID.GOOGLE_CALENDAR,
-                onSignIn: () => beginConsentSignIn(CONSENT_SERVICE_ID.GOOGLE_CALENDAR),
-                onRemoveAccount: removeCalendarAccount,
-                onToggleCalendar: toggleCalendarSelected,
-                onRefresh: () => act(ACT_KIND.CALENDAR_REFRESH),
-              },
-              appleCalendar: {
-                choices: appleCalendarObserved?.calendars ?? [],
-                held: slotHeldExcept(CONSENT_SERVICE_ID.APPLE_CALENDAR),
-                connecting:
-                  appleCalendarBusy ||
-                  consentConnect.entry?.serviceId === CONSENT_SERVICE_ID.APPLE_CALENDAR,
-                onSignIn: () => void connectAppleCalendar(),
-                onDisconnect: disconnectAppleCalendar,
-                onToggleCalendar: toggleAppleCalendarSelected,
-                revoked: appleCalendarObserved?.revoked === true,
-              },
-              linear: {
-                held: slotHeldExcept(CONSENT_SERVICE_ID.LINEAR),
-                connecting: consentConnect.entry?.serviceId === CONSENT_SERVICE_ID.LINEAR,
-                onSignIn: () => beginConsentSignIn(CONSENT_SERVICE_ID.LINEAR),
-                onDisconnect: disconnectLinear,
-              },
-              superset: (() => {
-                const supersetAgentDefault =
-                  settings.workspaceAgentDefaults?.[SUPERSET_WORKSPACE_PROVIDER_ID]?.agent;
-                const superset: SupersetControl = {
-                  installed: state.superset.installed,
-                  connected: supersetConnected,
-                  held: credentialHeld.current || consentConnectHeld.current,
-                  connecting: supersetSignInHeld.current,
-                  onConnect: beginSupersetSignIn,
-                  onDisconnect: disconnectSuperset,
-                  agents: [
-                    ...new Set(
-                      workspaceProjects
-                        .filter((project) => project.providerId === SUPERSET_WORKSPACE_PROVIDER_ID)
-                        .flatMap((project) => project.spawnableAgents ?? []),
-                    ),
-                  ],
-                  onDefaultAgentChange: changeSupersetAgentDefault,
-                };
-                if (supersetAgentDefault) {
-                  superset.defaultAgent = supersetAgentDefault;
-                }
-                return superset;
-              })(),
+              calendar: connections.calendar,
+              appleCalendar: connections.appleCalendar,
+              linear: connections.linear,
+              superset: connections.supersetControl,
               onQuit: () => tell(ACT_KIND.WINDOW_QUIT),
               shortcuts,
               searchOpen: settingsSearchOpen,
@@ -2522,10 +2122,10 @@ export function App(): React.JSX.Element {
           consent-connect pills split the remaining case by which entry
           holds the slot. A pill held through an old exit must not resurface
           under another's wait. */}
-      {signInWait === undefined ? (
+      {connections.signInWait === undefined ? (
         <>
           <KeySlot
-            control={credentials}
+            control={connections.credentials}
             source={slotSource}
             drawn={slotOpen && slotOccupant.current === PANEL_STAND_DOWN.KEY}
             measure={slotElement}
@@ -2533,19 +2133,10 @@ export function App(): React.JSX.Element {
           {/* The panel stood down while a consent sign-in waits on the
               browser, on the key slot's exact terms. */}
           <ConsentConnectSlot
-            entry={consentConnect.entry}
+            entry={connections.consentEntry}
             drawn={slotOpen && slotOccupant.current === PANEL_STAND_DOWN.CONSENT}
-            onCancel={cancelConsentSignIn}
-            onReopen={() => {
-              // The page reopened is the one the waiting flow built, named by
-              // the service the entry says is waiting — no address from here.
-              // A wait with no page to reopen has no action in its row, and the
-              // slot draws no button for it either.
-              const waiting = consentConnect.latest()?.serviceId;
-              if (!waiting) return;
-              const acts = CONSENT_ACTIONS[waiting];
-              if ("reopen" in acts) acts.reopen();
-            }}
+            onCancel={connections.cancelConsentSignIn}
+            onReopen={connections.reopenConsentPage}
             onOpenSystemSettings={() => tell(ACT_KIND.CALENDAR_OPEN_SETTINGS)}
             measure={connectElement}
           />
@@ -2555,20 +2146,20 @@ export function App(): React.JSX.Element {
               drawn={slotOpen && slotOccupant.current === PANEL_STAND_DOWN.SUPERSET}
               onSubmit={(code) => tell(ACT_KIND.SUPERSET_SUBMIT_CODE, { code })}
               onReopen={() => tell(ACT_KIND.SUPERSET_REOPEN_SIGN_IN)}
-              onCancel={cancelSupersetSignIn}
-              onRetry={beginSupersetSignIn}
+              onCancel={connections.cancelSupersetSignIn}
+              onRetry={connections.beginSupersetSignIn}
               onChooseOrganization={(slug) => tell(ACT_KIND.SUPERSET_CHOOSE_ORGANIZATION, { slug })}
               measure={connectElement}
             />
           ) : null}
         </>
       ) : null}
-      {credentialsEntry.entry === undefined && consentConnect.entry === undefined ? (
+      {connections.credentialEntry === undefined && connections.consentEntry === undefined ? (
         /* The panel stood down to the account sign-in it is waiting on. */
         <SignInSlot
-          {...(signInWait ? { provider: signInWait } : undefined)}
+          {...(connections.signInWait ? { provider: connections.signInWait } : undefined)}
           drawn={slotOpen}
-          onCancel={cancelSignIn}
+          onCancel={connections.cancelSignIn}
           measure={signInSlotElement}
         />
       ) : null}
